+++ cryptofeed/backends/redis.py
@@ -17,11 +17,16 @@ def trades_none_to_str(data):
 
 
 class RedisCallback(BackendQueue):
-    def __init__(self, host='127.0.0.1', port=6379, socket=None, key=None, numeric_type=float, **kwargs):
+    def __init__(self, host='127.0.0.1', port=6379, socket=None, key=None, numeric_type=float, writer_interval=0.01, **kwargs):
         """
         setting key lets you override the prefix on the
         key used in redis. The defaults are related to the data
         being stored, i.e. trade, funding, etc
+
+        writer_interval:
+        the frequency writer sleep when there is nothing in
+        data queue. 0 makes this thread consuming a lot of CPU.
+        while large interval put pressure on queue.
         """
         prefix = 'redis://'
         if socket:
@@ -32,6 +37,7 @@ class RedisCallback(BackendQueue):
         self.numeric_type = numeric_type
         self.running = True
         self.exited = False
+        self.writer_interval = writer_interval
 
     async def stop(self):
         self.running = False
@@ -59,7 +65,7 @@ class RedisZSetCallback(RedisCallback):
 
             count = self.queue.qsize()
             if count == 0:
-                await asyncio.sleep(0)
+                await asyncio.sleep(self.writer_interval)
             elif count > 1:
                 async with self.read_many_queue(count) as updates:
                     async with self.redis.pipeline(transaction=False) as pipe:
@@ -84,7 +90,7 @@ class RedisStreamCallback(RedisCallback):
 
             count = self.queue.qsize()
             if count == 0:
-                await asyncio.sleep(0)
+                await asyncio.sleep(self.writer_interval)
             elif count > 1:
                 async with self.read_many_queue(count) as updates:
                     async with self.redis.pipeline(transaction=False) as pipe: