+++ CHANGES.md
@@ -22,6 +22,7 @@
  * Feature: Type checking in Cython code (disabled by default, enable in setup.py)
  * Bugfix: Fix type issues in OKEx and Binance Futures - some numeric data being returned as string
  * Bugfix: Fix symbol normalization in FTX and Huoni Swap
+ * Feature: Redis backend to choose sleep interval for writer
 
 ### 2.0.0 (2021-09-11)
  * Feature: Binance REST support